application chain now works as before

+++ processFile.rb
@@ -8,24 +8,27 @@
 
 require 'pathname'
 require 'net/http'
+require 'nokogiri'
+require 'csv'
 
 ##
 ## Helper methods
 ##
 
-def get_full_year(filename)
-  partial_year = /^...(\d\d)\d\d\d\d\.zip$/.match(filename)[1]
+def get_date_fields(filename)
+  partial_year, month, day = /^...(\d\d)(\d\d)(\d\d)\./.match(filename)[1]
   full_year = (partial_year.to_i < 50 ? "20" : "19") + partial_year
+  return full_year, month, day
 end
 
 def extract_download_params(filename)
   download_server = "patents.reedtech.com"
 
   if filename =~ /^ipa\d{6}.zip$/  # application
-    full_year = get_full_year filename
+    full_year, month, day = get_date_fields filename
     server_path = "/downloads/ApplicationFullText/#{full_year}/#{filename}"
   elsif filename =~ /^ipg\d{6}.zip/ # grant
-    full_year = get_full_year filename
+    full_year, month, day = get_date_fields filename
     server_path = "/downloads/GrantRedBookText/#{full_year}/#{filename}"
   else
     raise "unknown file type (#{filename})"
@@ -48,8 +51,20 @@ def download_file(server, path, local_filename=nil)
   end
 end
 
-def line_contains_nsf(line)
-  /\bN\W*S\W*F\b|National Science Foundation/.match(line)
+def text_contains_nsf_term(line)
+  /\bN\W*S\W*F\b|National Science Foundation/m.match(line)
+end
+
+def extract_govt_interest(text)
+  processinstr1 = '<?federal-research-statement description="Federal Research Statement" end="lead"?>'
+  processinstr2 = '<?federal-research-statement description="Federal Research Statement" end="tail"?>'
+  matches = text.match(/#{Regexp.escape(processinstr1)}(.*)#{Regexp.escape(processinstr2)}/m)
+  matches[1].strip if matches
+end
+
+def block_has_nsf_govt_interest(lines)
+  gov_int = extract_govt_interest(lines.join)
+  text_contains_nsf_term(gov_int)
 end
 
 def extract_file(xml_filename, extract_filename)
@@ -64,7 +79,6 @@ def extract_file(xml_filename, extract_filename)
       line_iter = fin.each_line
       current_block_lines = []
       inside_block        = false
-      block_contains_nsf  = false
 
       line_count  = 0
       block_count = 0
@@ -78,28 +92,21 @@ def extract_file(xml_filename, extract_filename)
         if inside_block
           current_block_lines << current_line
 
-          match_info = line_contains_nsf(current_line)
-          if match_info
-            block_contains_nsf = true
-            puts "line #{line_count} in block #{block_count} contains NSF-related term (#{match_info})"
-          end
-
           if current_line =~ %r{</us-patent-application>} # leaving the block
             #puts "line #{line_count}: leaving block"
-            if block_contains_nsf
+            if block_has_nsf_govt_interest(current_block_lines)
               #puts "block #{block_count} contains NSF-related term"
               current_block_lines.each{|line| fout.write line}
             end
             current_block_lines = []
             inside_block = false
           end
-        elsif current_line =~ %r{<us-patent-application .*>} # entering a block
+        elsif current_line =~ %r{<us-patent-application.*>} # entering a block
           #puts "line #{line_count}: entering block"
           current_block_lines << current_line
 
           inside_block        = true
           block_count        += 1
-          block_contains_nsf  = false
         end
       end while !fin.eof?
     end
@@ -109,17 +116,143 @@ def extract_file(xml_filename, extract_filename)
   end
 end
 
+## To use this class, pass it a block which takes Nokogiri::XML node (or whatever it's called)
+## and return something which can be converted to a string
+class Extractor
+  def initialize(field_name, &extractor)
+    @field_name = field_name
+    @extractor  = extractor
+  end
+
+  def field_name
+    @field_name
+  end
+
+  def process(xml, filename)
+    @extractor.call(xml, filename).to_s
+  end
+end
+
+def produce_applications_report(extract_filename, report_filename)
+  return unless extract_filename =~ /ipa/
+
+  ##
+  ## Create a list of Extractors
+  ##
+
+  #FIELDS = %w{appno pubdate pubnum title abstract invs assignee xref filedate govint parentcase childcase date371 pctpubno}
+
+  extractors = []
+
+  extractors << Extractor.new("appno") do |app, filename|
+    app.xpath(".//application-reference/document-id/doc-number/text()").to_s
+  end
+  extractors << Extractor.new("pubdate") do |app, filename|
+    app.xpath(".//publication-reference/document-id/date/text()").to_s
+  end
+  extractors << Extractor.new("pubnum") do |app, filename|
+    app.xpath("./us-bibliographic-data-application/publication-reference/document-id/doc-number/text()").to_s
+  end
+  extractors << Extractor.new("title") do |app, filename|
+    app.xpath('.//invention-title/text()').to_s
+  end
+  extractors << Extractor.new("abstract") do |app, filename|
+    app.xpath(".//abstract/p/text()").to_s
+  end
+  extractors << Extractor.new("invs") do |app, filename|
+    full_year, month, day = get_date_fields(filename)
+    inventor_xpath = (full_year.to_i < 2007) ? './/applicant' : './/inventor/addressbook'
+    inventors = app.xpath(inventor_xpath).collect do |inventor|
+      inventor.xpath('.//first-name/text()').to_s + " " + inventor.xpath('.//last-name/text()').to_s
+    end
+    inventors.map{|i| "[#{i}]"}.join
+  end
+  extractors << Extractor.new("assignee") do |app, filename|
+    full_year, month, day = get_date_fields(filename)
+    case full_year.to_i
+    when 2012..2014
+      assignees = app.xpath('.//assignees/assignee').collect do |assignee|
+        assignee.xpath('./addressbook/orgname/text()').to_s
+      end
+      assignees.map{|a| "[#{a}]"}.join
+    end
+  end
+  extractors << Extractor.new("xref") do |app, filename|
+    if app.at_xpath('//processing-instruction("cross-reference-to-related-applications")')
+      processxref1 = '<?cross-reference-to-related-applications description="Cross Reference To Related Applications" end="lead"?>'
+      processxref2 = '<?cross-reference-to-related-applications description="Cross Reference To Related Applications" end="tail"?>'
+      matches = app.to_s.match(/#{Regexp.escape(processxref1)}(.*)#{Regexp.escape(processxref2)}/m)
+      if matches
+        Nokogiri::XML.fragment(matches[1].strip).xpath("./p/text()").to_s
+      end
+    else
+      app.xpath(".//description/heading[contains(.,'CROSS-REFERENCE') or contains(.,'CROSSREF') or contains(.,'CROSS REFERENCE')]").to_s
+    end
+  end
+  extractors << Extractor.new("filedate") do |app, filename|
+    app.xpath(".//application-reference/document-id/date/text()").to_s
+  end
+  extractors << Extractor.new("govint") do |app, filename|
+    Nokogiri::XML.fragment(extract_govt_interest(app.to_s)).xpath("./p/text()").to_s
+  end
+  extractors << Extractor.new("parentcase") do |app, filename|
+    app.xpath(".//us-related-documents//parent-doc/document-id/doc-number/text()").first.to_s
+  end
+  extractors << Extractor.new("childcase") do |app, filename|
+    app.xpath(".//us-related-documents//child-doc/document-id/doc-number/text()").first.to_s
+  end
+  extractors << Extractor.new("date371") do |app, filename|
+    app.xpath(".//us-371c124-date/date/text()").to_s
+  end
+  extractors << Extractor.new("pctpubno") do |app, filename|
+    app.xpath(".//pct-or-regional-filing-data/document-id/doc-number/text()").to_s
+  end
+
+  ##
+  ## Run the Extractors against the file
+  ##
+
+  all_extracts = []
+  File.open(extract_filename, "r") do |fin|
+    doc = Nokogiri::XML(fin)
+
+    doc.xpath('.//us-patent-application').each do |app|
+      # Check that there is a Federal Research Statement. If so, continue, if not, jump to next patent app
+      next unless doc.at_xpath('//processing-instruction("federal-research-statement")')
+
+      app_extracts = extractors.collect{|e| e.process(app, extract_filename)}
+      all_extracts << app_extracts
+    end
+  end
+
+  ##
+  ## Write the output report
+  ##
+
+  CSV.open(report_filename,"w") do |csv|
+    csv << extractors.collect{|e| e.field_name}
+    all_extracts.each{|app_extract| csv << app_extract}
+  end
+
+end
+
+def produce_grants_report(extract_filename, report_filename)
+  return unless extract_filename =~ /ipg/
+
+end
+
 ##
 ## Parse command-line arguments
 ##
 
-actions   = ARGV.select{|s| s =~ /^download|unzip|extract|cleanup$/}
+actions   = ARGV.select{|s| s =~ /^download|unzip|extract|report|cleanup$/}
 filenames = ARGV - actions
 filenames = filenames.map{|f| f.gsub(/\..*$/, "")}
 
 should_download = actions.empty? || (actions.include? "download")
 should_unzip    = actions.empty? || (actions.include? "unzip")
 should_extract  = actions.empty? || (actions.include? "extract")
+should_report   = actions.empty? || (actions.include? "report")
 should_cleanup  = actions.empty? || (actions.include? "cleanup")
 
 puts "actions   = #{actions}"
@@ -134,9 +267,11 @@ filenames.each do |filename|
 
   begin
 
-    zip_filename      = "#{filename}.zip"
-    xml_filename      = "#{filename}.xml"
-    extract_filename  = "#{filename}.extract"
+    zip_filename            = "#{filename}.zip"
+    xml_filename            = "#{filename}.xml"
+    extract_filename        = "#{filename}.extract"
+    report_apps_filename    = "#{filename}.apps"
+    report_grants_filename  = "#{filename}.grants"
 
     if should_download
       puts "  downloading #{filename}"
@@ -159,12 +294,21 @@ filenames.each do |filename|
       extract_file xml_filename, extract_filename
     end
 
+    if should_report
+      puts "  reporting on #{filename}"
+
+      raise "#{extract_filename} doesn't exist" unless File.exists? extract_filename
+      produce_applications_report extract_filename, report_apps_filename
+      produce_grants_report       extract_filename, report_grants_filename
+    end
+
     if should_cleanup
       puts "  cleaning up #{filename}"
 
       File.delete zip_filename
       File.delete xml_filename
       File.delete extract_filename
+      # don't delete report files!
     end
 
   rescue StandardError => e
